chore(off-chain_sync): renames file
